refactor: eliminate Server god-struct — move handlers to handler groups
WS1: Policy handlers → PolicyHandlers struct
- Move 10 methods from policy_handler.go off *Server to *PolicyHandlers
- Convert querySchemaNodes/querySchemaEdges to standalone functions
- Unify schemaPool/policyPool into shared tenantPool() helper
- Wire via SchemaHandlers.Policy

WS2: Schema graph handlers → SchemaGraphHandlers struct
- Move GetSchemaGraph, GetTableRelationships to *SchemaGraphHandlers
- Wire via SchemaHandlers.Graph

WS3: Remaining route handlers → handler groups
- Move handleGetTables/handleGetTableSchema/handleGetSchemas/
  handleExecuteQuery/handleRefreshSchema → SchemaAdminHandlers
- Move handleRealtimeStats → RealtimeHandlers.HandleStats
- Move InvalidateSchemaCache → SchemaAdminHandlers

WS4: Dissolve getters, loaders, dead code
- Remove GetTenantConfigLoader (zero callers)
- Remove SchemaCache/GetWebhookTriggerService/GetLoggingService getters
  (callers use handler group fields directly)
- Move LoadFunctionsFromFilesystem/LoadJobsFromFilesystem to main.go
  (call handler group methods directly)
- Remove LoadAIChatbotsFromFilesystem (test-only, removed)

Result: Server methods reduced from 37 → 8 (lifecycle + 3 remaining accessors)
