+++ internal/api/routes_admin.go
@@ -15,10 +15,10 @@ func (s *Server) buildAdminRouteDeps() *routes.AdminDeps {
 		// Subgroup dependencies
 		Branch: s.buildBranchRouteDeps(),
 		Schema: &routes.SchemaAdminDeps{
-			GetTables:               s.handleGetTables,
-			GetTableSchema:          s.handleGetTableSchema,
-			GetSchemas:              s.handleGetSchemas,
-			ExecuteQuery:            s.handleExecuteQuery,
+			GetTables:               s.Schema.Admin.GetTables,
+			GetTableSchema:          s.Schema.Admin.GetTableSchema,
+			GetSchemas:              s.Schema.Admin.GetSchemas,
+			ExecuteQuery:            s.Schema.Admin.ExecuteQuery,
 			ListSchemasDDL:          s.Schema.DDL.ListSchemas,
 			CreateSchemaDDL:         s.Schema.DDL.CreateSchema,
 			ListTablesDDL:           s.Schema.DDL.ListTables,
@@ -34,18 +34,18 @@ func (s *Server) buildAdminRouteDeps() *routes.AdminDeps {
 			DisableRealtime:         s.Realtime.Admin.HandleDisableRealtime,
 			ExecuteSQL:              s.sqlHandler.ExecuteSQL,
 			ExportTypeScript:        s.Schema.Export.HandleExportTypeScript,
-			RefreshSchema:           s.handleRefreshSchema,
-			GetSchemaGraph:          s.GetSchemaGraph,
-			GetTableRelationships:   s.GetTableRelationships,
-			GetTablesWithRLS:        s.GetTablesWithRLS,
-			GetTableRLSStatus:       s.GetTableRLSStatus,
-			ToggleTableRLS:          s.ToggleTableRLS,
-			ListPolicies:            s.ListPolicies,
-			CreatePolicy:            s.CreatePolicy,
-			UpdatePolicy:            s.UpdatePolicy,
-			DeletePolicy:            s.DeletePolicy,
-			GetPolicyTemplates:      s.GetPolicyTemplates,
-			GetSecurityWarnings:     s.GetSecurityWarnings,
+			RefreshSchema:           s.Schema.Admin.RefreshSchema,
+			GetSchemaGraph:          s.Schema.Graph.GetSchemaGraph,
+			GetTableRelationships:   s.Schema.Graph.GetTableRelationships,
+			GetTablesWithRLS:        s.Schema.Policy.GetTablesWithRLS,
+			GetTableRLSStatus:       s.Schema.Policy.GetTableRLSStatus,
+			ToggleTableRLS:          s.Schema.Policy.ToggleTableRLS,
+			ListPolicies:            s.Schema.Policy.ListPolicies,
+			CreatePolicy:            s.Schema.Policy.CreatePolicy,
+			UpdatePolicy:            s.Schema.Policy.UpdatePolicy,
+			DeletePolicy:            s.Schema.Policy.DeletePolicy,
+			GetPolicyTemplates:      s.Schema.Policy.GetPolicyTemplates,
+			GetSecurityWarnings:     s.Schema.Policy.GetSecurityWarnings,
 			DumpInternalSchema:      s.Schema.InternalSchema.DumpSchema,
 			PlanInternalSchema:      s.Schema.InternalSchema.PlanSchema,
 			ApplyInternalSchema:     s.Schema.InternalSchema.ApplySchema,
